+++ ui/src/app/modules/blog/components/blog-post-view/blog-post-view.component.ts
@@ -67,6 +67,8 @@ export class BlogPostViewComponent implements OnInit {
     }
 
     deletePost(id: number): void {
+        if(!window.confirm(`Are you sure you want to delete this post?`)) return;
+
         this.blogApiService.deletePost(id).subscribe((res: any) => {
             this.notificationService.createNotification('Successfully deleted blog post!');
             this.router.navigate(['/blog']);
