+++ ui/src/app/modules/blog/components/blog-view/blog-view.component.ts
@@ -94,6 +94,8 @@ export class BlogViewComponent implements OnInit {
     }
 
     deleteTopic(topic: BlogTopic): void {
+        if(!window.confirm(`Are you sure you want to delete this topic?`)) return;
+
         this.blogApiService.deleteTopic(topic.id).subscribe((res: any) => {
             this.removeTopic(topic.id);
             this.notificationService.createNotification('Successfully deleted blog topic!');
